requirements.txt
# Requirements for Agentic Explorations
python-dotenv
openai
